Make beginning file
Intentions are that the script will always have at least one argument
for the 'mode' to use. Arguments to be handled later, even though the
test for their existance is made now.

The parsing of the CLI arguments will get much larger, more complex,
and eventually require l10n workups.

Baseline test of does the script execute and respond to the missing
argument?

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>