Move messages to sourced file
To improve the process of i18n and l10n, the messages need to be in a
file independent of the operations. The actual creation of the files
for LocalData will be done much later. As there are no translators at
hand, it will be empty initially, but "built" for if/when there is the
ability to have translations made.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>