Relocate messages data file
Although not yet implemented, the messages should be in a format able
to be used for i18n/l10n. I also don't like filling the project's main
directory with lots of files.

This is the beginning of a refactoring to clean up the project root and
make something like a CleanCode version of the script. Bash does not,
in a practical sense, support the full encapsulation of C++, or Java,
and I'm not going to attempt to localize all the data. I will still
keep the processing of data in "modules." In a shell script using a
bunch of functions, and files, is only asking for performance issues.
In this case, with downloads from the Internet and multi-gig data
being written to USB sticks, the performace bottleneck won't be the
script.

Future, or final, version will probably un-refactor it all anyway.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>