Force root permissions
It is not needed, yet, by the script. However, in time the whole
point is writing images directly to USB sticks at the low-level,
as in `/dev/sdx` and the rest of the data should be under root
control. Add the piece to force it now and it'll be there when the
script gets to that point.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>